refactor(config_routes): extract session.py + monitoring.py from _factory.py (3/8)
Schritt 3 von 8 im Split-Plan
(`docs/superpowers/plans/2026-04-29-config-routes-split.md`).

What moved
- `_register_session_routes()` (~308 LOC) -> `cognithor.channels.config_routes.session`. Vault, Session-Isolation, Workspace-Isolation, Multi-Tenancy, Chat-History.
- `_register_memory_routes()` (~200 LOC) -> dieselbe Datei `session.py`. Memory-Hygiene, Integrity-Checker, Decision-Explainer / Explainability — passt thematisch (beides Per-Session-Lebenszeit-Daten).
- `_register_monitoring_routes()` (~165 LOC) -> `cognithor.channels.config_routes.monitoring`. Live-Metriken, Audit, SSE-Stream, Heartbeat-Status.
- `_register_prometheus_routes()` (~45 LOC) -> dieselbe Datei `monitoring.py`. `/metrics` Prometheus Exposition Format. Teilt sich `get_hub`-Closure mit Monitoring.

`_factory.py` schrumpft von 5 503 -> 4 794 Zeilen (-13 %).

Per-Sub-Module-Imports
- `session.py`: `Request`, `HTTPException`, `Any`, plus internal `log`. Lazy-Imports innerhalb der Handler bleiben verbatim (`MultiUserIsolation`, `WorkspaceGuard` usw.).
- `monitoring.py`: `HTTPException`, `Any`, `log`. `Request` wird hier nicht gebraucht. `config_manager: Any = None` als Parameter — daher kein `ConfigManager`-Type-Import noetig.
F401-clean.

Aufruf-Reihenfolge
`create_config_routes()` ruft `_register_session_routes`, `_register_memory_routes`, `_register_monitoring_routes`, `_register_prometheus_routes` exakt in der bestehenden Position der Sequenz weiter auf. FastAPI Path-Match-Reihenfolge unveraendert. `_get_hub`-Closure bleibt in `_factory.py` und wird wie zuvor als Callable an die zwei Monitoring-Helper uebergeben.

Verification
- `pytest test_route_inventory_unchanged` -> passed (303/303 endpoints).
- Konsumenten-Sweep ueber 11 Test-Files -> 396 passed.
- `ruff check src/cognithor/channels/config_routes/` -> All checks passed.
- `ruff format --check` -> 6 files already formatted.
